Assign lang upon the creation of the testing element
This fixes issue 5. Firefox appears to have more fine-grained control on
font configuration, which makes it return surprisng values ini non-English
locales. Before this change the default font in the test suite and in
the actual element would be different if the OS locale wasn't English.
Hence, it's better to initialize the lang before initializing
tests and webfonts.