Fix wrong language comparison
If the applied element has no language attribute we use inherited
language value(present as this.language). But in this place we were
using elements lang attribute directly. Which can be undefined sometimes
and can cause expliticit fontfamily style added to child elements

Eg:
<html lang='en'>
..
<body>
<div lang='en'>

After applying FontA to body should be
<html lang='en'>
..
<body style="fontfamily='FontA, fallbacks';">
<div lang='en'>

but we were getting
<html lang='en'>
..
<body style="fontfamily='FontA, fallbacks';">
<div lang='en' style="fontfamily='FontA, fallbacks';>

That makes resetting not applying to child div since we never touch
child elements with explicit fontfamily value