Remove magic date from version query string
The '&20120101' component of the font URL version suffix traces back to the
first substantial code commit, 1ac3aa5ef6. I suspect it is a stray bit of debug
/ dev code that became part of the furniture after a while. Since it is fixed,
it is not causing a lot of harm, but it makes it difficult to apprehend the
browser cache invalidation strategy being used.